requirements.txt
api==0.0.7
arrow==0.13.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.184
awsebcli==3.15.2
backcall==0.1.0
beautifulsoup4==4.7.1
binaryornot==0.4.4
blessed==1.15.0
boto==2.49.0
botocore==1.12.174
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cookiecutter==1.6.0
crypto==1.4.1
cryptography==2.6.1
decorator==4.4.0
Django==2.2.6
django-cors-headers==2.5.3
django-environ==0.4.5
django-filter==2.1.0
django-hitcount==1.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-scheduler==0.8.8
django-storages==1.7.1
django-taggit==1.1.0
django-test-plus==1.2.0
django-uuidfield==0.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
environ==1.0
flake8==3.7.7
future==0.16.0
gunicorn==20.0.0
icalendar==4.0.3
idna==2.7
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==2.6.0
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
Naked==0.1.31
ndg-httpsclient==0.5.1
nose==1.3.7
npm==0.1.1
numpy==1.16.4
optional-django==0.1.0
pandas==0.24.2
parso==0.4.0
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
pluggy==0.9.0
poyo==0.4.2
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pusher==2.1.3
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytest==4.4.1
pytest-django==3.4.8
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.1
selenium==3.141.0
semantic-version==2.5.0
Serializer==0.2.1
shellescape==3.4.1
simplejson==3.16.0
six==1.11.0
soupsieve==1.9.1
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
urllib3==1.25.6
uWSGI==2.0.18
views==0.3
virtualenv==16.3.0
virtualenv-clone==0.5.1
wcwidth==0.1.7
websocket-client==0.56.0
whichcraft==0.5.2
